Restructuring tangle abstraction and fixing errors (#1050)
* Restructuring and fix errors

* wip
