--- src/data/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2023 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! Defines types that allow for unified data processing.
-
-mod input_source;
-
-pub use input_source::*;
