+++ src/tangle/sources/inx.rs
@@ -1,0 +1,2 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
